# Use PEP 393 new APIs