markupsafe doesn't support Python 3.2 anymore.
So just use new API at Python 3.