@ThiefMaster 2017-02-12
Small recommendation: Don't mix commits changing just whitespace with other commits