@methane 2017-02-12
Python 2:
```
before:
=============================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           100000 loops, best of 3: 5.01 usec per loop
largestring                     1000 loops, best of 3: 372 usec per loop
long_empty_string               10000 loops, best of 3: 81.4 usec per loop
long_suffix                     1000 loops, best of 3: 977 usec per loop
short_empty_string              100000 loops, best of 3: 4.57 usec per loop
--------------------------------------------------------------------------------

patched:
===============================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           100000 loops, best of 3: 2.02 usec per loop
largestring                     1000 loops, best of 3: 347 usec per loop
long_empty_string               10000 loops, best of 3: 58.2 usec per loop
long_suffix                     1000 loops, best of 3: 860 usec per loop
short_empty_string              1000000 loops, best of 3: 1.72 usec per loop
--------------------------------------------------------------------------------
```