@methane 2017-02-12
Python 3
```
before
==============================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           100000 loops, best of 3: 2.44 usec per loop
largestring                     1000 loops, best of 3: 262 usec per loop
long_empty_string               100000 loops, best of 3: 15.4 usec per loop
long_suffix                     1000 loops, best of 3: 505 usec per loop
short_empty_string              100000 loops, best of 3: 2.18 usec per loop
--------------------------------------------------------------------------------

patched
==============================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           100000 loops, best of 3: 2.26 usec per loop
largestring                     10000 loops, best of 3: 123 usec per loop
long_empty_string               100000 loops, best of 3: 13.5 usec per loop
long_suffix                     1000 loops, best of 3: 322 usec per loop
short_empty_string              100000 loops, best of 3: 2.07 usec per loop
--------------------------------------------------------------------------------
```