@methane 2018-04-27
```
before:

================================================================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           200000 loops, best of 5: 1.91 usec per loop
largestring                     2000 loops, best of 5: 199 usec per loop
long_empty_string               20000 loops, best of 5: 15.9 usec per loop
long_suffix                     1000 loops, best of 5: 386 usec per loop
short_empty_string              200000 loops, best of 5: 1.66 usec per loop
--------------------------------------------------------------------------------

patched:

================================================================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           200000 loops, best of 5: 1.34 usec per loop
largestring                     2000 loops, best of 5: 111 usec per loop
long_empty_string               20000 loops, best of 5: 10.2 usec per loop
long_suffix                     1000 loops, best of 5: 234 usec per loop
short_empty_string              200000 loops, best of 5: 1.21 usec per loop
--------------------------------------------------------------------------------
```