@davidism 2018-04-27
We no longer support 2.6 and 3.3 either. Would that affect this code (probably make it simpler)?