@methane 2018-04-27
I don't think so.

On the other hand, I can remove Python 2 specific code and share same logic (GET_DELAT and DO_ESCAPE macro).