@davidism 2018-04-27
@methane I'm not very familiar with C / the C-API. If you have some time, would you mind checking #74 just so I know it makes sense?