@methane 2018-04-27
Python 2.7.13

```
before:
================================================================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           1000000 loops, best of 3: 1.79 usec per loop
largestring                     1000 loops, best of 3: 261 usec per loop
long_empty_string               10000 loops, best of 3: 42.7 usec per loop
long_suffix                     1000 loops, best of 3: 630 usec per loop
short_empty_string              1000000 loops, best of 3: 1.68 usec per loop
--------------------------------------------------------------------------------

after:
================================================================================
Running benchmark for MarkupSafe
--------------------------------------------------------------------------------
basic                           1000000 loops, best of 3: 1.79 usec per loop
largestring                     1000 loops, best of 3: 286 usec per loop
long_empty_string               10000 loops, best of 3: 38.4 usec per loop
long_suffix                     1000 loops, best of 3: 726 usec per loop
short_empty_string              1000000 loops, best of 3: 1.46 usec per loop
--------------------------------------------------------------------------------
```

Hmm, 379d0d0 makes escaping little slower on PY2.7.