@davidism 2018-04-27
I'm fine with reverting it if it makes it faster.