@methane 2018-04-28
I removed last commit.
We can remove PY2 only code after PY2 EOL anyway.  So keeping some PY2 only code for few years is not big problem.