@ThiefMaster 2018-04-28
> We can remove PY2 only code after PY2 EOL anywa

More like "only after EOL being dropped by all major distributions, and not being relevant anymore" ;)