@methane 2018-04-30
"after" didn't mean "right after".
I don't care when you drop PY2 support.

(FYI, Django 2.0 has significant performance gain from dropping PY2 support. 😉 )