@davidism 2018-05-04
Rebased to master. Converted to pytest.